frontend/public/image/gamesgroup (1)/gamesgroup/requirements.txt
amqp==5.1.1
ansicon==1.89.0
arrow==1.1.1
asgiref==3.3.4
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
blessed==1.18.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.11
django-allauth==0.51.0
django-auto-logout==0.5.0
django-clearcache==1.2.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-decouple==2.1
django-phone-field==1.8.1
django-picklefield==3.0.1
django-redshift-backend==3.0.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.4.1
drf-yasg==1.21.4
exceptiongroup==1.0.3
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==3.4
importlib==1.0.4
importlib-resources==5.10.1
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jinxed==1.1.0
jsonschema==4.17.3
kombu==5.2.4
MarkupSafe==2.1.1
netbox-api==0.4.3
netbox-secretstore==1.3.0
numpy==1.23.4
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
pandas==1.5.1
petl==1.7.11
Pillow==9.2.0
ping3==4.0.3
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
psycopg2==2.9.4
pycparser==2.21
pycryptodome==3.16.0
PyJWT==1.7.1
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
six==1.16.0
SQLAlchemy==1.4.42
sqlparse==0.4.1
tabulate==0.9.0
tomli==2.0.1
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.11.0
